+++ src/org/sosy_lab/cpachecker/cpa/smg/join/SMGJoinTargetObjects.java
@@ -225,7 +225,7 @@ final class SMGJoinTargetObjects {
 
     destSMG.removeObjectAndEdges(targetObject);
 
-    Set<Integer> restricted = new HashSet<>();
+    Set<Long> restricted = new HashSet<>();
 
     switch (targetObject.getKind()) {
       case DLL:
@@ -242,7 +242,8 @@ final class SMGJoinTargetObjects {
     for (SMGEdgeHasValue hve : hves) {
       Integer val = hve.getValue();
 
-      if (!restricted.contains(val) && val != 0) {
+      //FIXME: require to identify why offsets are mixed with values
+      if (!restricted.contains(val.longValue()) && val != 0) {
 
         if (destSMG.isPointer(val)) {
           SMGObject reachedObject = destSMG.getPointer(val).getObject();
