+++ src/org/sosy_lab/cpachecker/util/predicates/pathformula/pointeraliasing/CToFormulaConverterWithPointerAliasing.java
@@ -526,7 +526,7 @@ public class CToFormulaConverterWithPointerAliasing extends CtoFormulaConverter 
       assert compositeType.getKind() != ComplexTypeKind.ENUM : "Enums are not composite: " + compositeType;
       for (final CCompositeTypeMemberDeclaration memberDeclaration : compositeType.getMembers()) {
         final String memberName = memberDeclaration.getName();
-        final int offset = typeHandler.getBitOffset(compositeType, memberName);
+        final long offset = typeHandler.getBitOffset(compositeType, memberName);
         final CType memberType = typeHandler.getSimplifiedType(memberDeclaration);
         final String newBaseName = getFieldAccessName(baseName, memberDeclaration);
         if (hasIndex(newBaseName, memberType, ssa) &&
